Add more error handling to prevent crashing
* width and height are now consistent uint16_t type
* Fixed most of the warnings shows in Visual Studio

+++ framework/include/sr/sprite.h
@@ -6,11 +6,11 @@
 
 typedef struct sr_sprite_p* sr_sprite;
 
-sr_sprite sr_sprite_create(uint32_t width, uint32_t height);
+sr_sprite sr_sprite_create(uint16_t width, uint16_t height);
 void sr_sprite_destroy(sr_sprite sprite);
 
-struct sr_pixel sr_sprite_get_pixel(sr_sprite sprite, uint32_t x, uint32_t y);
-void sr_sprite_set_pixel(sr_sprite sprite, uint32_t x, uint32_t y, struct sr_pixel pixel);
+struct sr_pixel sr_sprite_get_pixel(sr_sprite sprite, uint16_t x, uint16_t y);
+void sr_sprite_set_pixel(sr_sprite sprite, uint16_t x, uint16_t y, struct sr_pixel pixel);
 struct sr_pixel* sr_sprite_get_data(sr_sprite sprite);
 
 #endif
