+++ framework/src/sprite_private.h
@@ -5,8 +5,8 @@
 
 struct sr_sprite_p
 {
-    uint32_t width;
-    uint32_t height;
+    uint16_t width;
+    uint16_t height;
     struct sr_pixel* data;
 };
